Localgov paragraphs accordion enhancements (#192)
* feat: adds two additional localgov_accordion paragraph fields:

1. Allow Multiple Open: when checked allows multiple panes to be open
   simultaneously
2. Display Show/Hide All: when checked, makes a new button available to
   expand or collapse all accordion panes

* feat: adds new localgov_subsites_paragraphs.install file

Implements hook_update_N() to add config for newly-added fields.

* feat: removes original .js file, renames .es6.js to .js

We're no longer writing es6 and transpiling to es5--now using es6
directly.

* feat: moves expand/collapse logic to dedicated functions

This is needed for forthcoming expand/collapse all function.

* feat: enable "allow multiple" behaviour

With this change, multiple panes can be opened simultaneously.

* feat: update field description to more correctly describe field behavior

* feat: provide markup for show/hide all button for accordions

* feat: adds, uses new expand/collapse fns, adds, uses new toggleAll() fn

* feat: adds minimal toggle styles; intelligently toggles button state

The show/hide all button will now adapt its behavior when e.g. set to
"Show all" and the last pane is expanded, or when e.g. set to "Hide all"
and the last pane is collapsed.

* feat: adds third parameter to in_array() calls per test fail

* feat: accounts for auto-allow-multiple when display show/hide all set

The javascript already enforces this, but it can only do so if the data
attributes are set correctly. This change ensures that they are.

* feat: updates weights of existing fields to show new checkboxes near top

* feat: replaces Yaml-sourced config with literal config

* feat: removes unused 'use', unused variable per coding-standards check